% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_errors_discoveries.R
\name{calc_errs_new}
\alias{calc_errs_new}
\title{Calculate the error and success proportions of tests for a single iteration}
\usage{
calc_errs_new(
  testobj,
  truevar_name,
  trueeffect_tol = .Machine$double.eps,
  blockid = "bF",
  thealpha = 0.05,
  fwer = FALSE,
  return_details = FALSE
)
}
\arguments{
\item{testobj}{Is an object arising from \link[manytestsr:find_blocks]{manytestsr::find_blocks} or
\link{adjust_block_tests}. It will contain block-level results.}

\item{truevar_name}{Is a string indicating the name of the variable
containing the true underlying causal effect (at the block level).}

\item{trueeffect_tol}{Is the smallest effect size below which we consider
the effect to be zero (by default is it floating point zero).}

\item{blockid}{A character name of the column in idat and bdat indicating the block.}

\item{thealpha}{Is the error rate for a given test (for cases where alphafn
is NULL, or the starting alpha for alphafn not null)}

\item{fwer}{Indicates that we are trying to control FWER. Right now, we do
this by default in report_detections but this indicates when we should be
looking at FDR}

\item{return_details}{TRUE means that the function should return a list of
the original data, a summary of the results,  a node level dataset, and a
copy of the original object that was provided as input. Default here is
FALSE. Only use TRUE when not using simulations.}
}
\value{
Most basically this function returns a single vector (or data.table with 1
row) with a variety of summaries of the performance of the testing procedure
across the blocks (i.e. leaves) and nodes (i.e. groups of blocks) in the
design.
}
\description{
This function takes output from \link[manytestsr:find_blocks]{manytestsr::find_blocks} or an equivalent
bottom-up testing function such as \link{adjust_block_tests} and returns the
proportions of errors and discoveries made across the multiple tests. To use
this function the input to \link[manytestsr:find_blocks]{manytestsr::find_blocks} must include a column
containing a true block-level effect in \code{truevar_name}. Repeated uses of
this function allow us to assess false discovery rates and family wise error
rates among other metrics of testing success.
}
