% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_test_sims.R
\name{draw_node_p_value}
\alias{draw_node_p_value}
\title{Draw a Single Node's P-value under Monotonicity}
\usage{
draw_node_p_value(is_alt, parent_p, beta_params)
}
\arguments{
\item{is_alt}{Logical; \code{TRUE} if the node is non-null, \code{FALSE} if null.}

\item{parent_p}{Numeric in \eqn{[0,1]}; the parent's final p-value.}

\item{beta_params}{Numeric vector of length 2, e.g. \code{c(a, b)}, specifying
the Beta(\code{a},\code{b}) distribution for alternative nodes. The actual
p-value is then scaled to \eqn{[parent\_p, 1]}.}
}
\value{
A numeric p-value in \eqn{[parent\_p, 1]}.
}
\description{
Generates a final p-value for a node given its parent's final p-value,
respecting the monotonic rule \eqn{p_\text{child} \ge p_\text{parent}}.
}
\details{
If \code{is_alt=FALSE}, we draw \eqn{\mathrm{Uniform}(parent\_p, 1)}.
If \code{is_alt=TRUE}, we draw \eqn{\mathrm{Beta}(a,b)} and linearly
rescale the result into \eqn{[parent\_p, 1]}.
}
\examples{
# Null node p-value if parent_p=0.05 all should be greater than or equal to .05
draw_node_p_value(FALSE, parent_p = 0.05, beta_params = c(0.1, 1))

# Alt node p-value if parent_p=0.05 all should be greater than or equal to .05
draw_node_p_value(TRUE, parent_p = 0.05, beta_params = c(0.1, 1))

}
