% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_test_sims.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Return a vector of all ancestors of a node}
\usage{
get_ancestors(node_id, k)
}
\arguments{
\item{node_id}{Integer; The global node id}

\item{k}{Integer; The number of nodes at each level of the k-ary tree}
}
\value{
a vector with the global ids of the ancestors of the node (always includes node 1 or the root node)
}
\description{
Given a 1-based global node id and information about the number of nodes at a level, return the global indices of all ancestors of that node.
}
\examples{
## This next returns c(3,1)
get_ancestors(node_id = 8, k = 3)
## This next returns c(4,2,1)
get_ancestors(node_id = 8, k = 2)

}
