% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_adjusters.R
\name{local_hommel_all_ps}
\alias{local_hommel_all_ps}
\title{Compute local Hommel p-value for a Vector of Child p-values}
\usage{
local_hommel_all_ps(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha (not used in this function)}
}
\value{
A vector of adjusted p-values
}
\description{
Given \eqn{k} child p-values, computes the Hommell adjusted p-values
}
\details{
TODO The Hommel adjustment and the hommel package is faster than p.adjust().
}
\examples{
local_simes(c(0.01, 0.04, 0.10, 0.20))

}
