% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_adjusters.R
\name{local_simes}
\alias{local_simes}
\title{Compute local Simes p-value for a Vector of Child p-values}
\usage{
local_simes(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha (not used in this function)}
}
\value{
A single numeric value: the Simes combination p-value.
}
\description{
Given \eqn{k} child p-values, computes the Simes p-value
\eqn{\min_{i=1\ldots k} \{ (k/i) * p_{(i)} \}}, where \eqn{p_{(1)} \le \ldots \le p_{(k)}}.
}
\details{
The Simes p-value is a valid test for the intersection hypothesis
\eqn{H_{\cap} : \text{all child hypotheses are null}}, under certain independence
or positive dependence conditions among the p-values.
}
\examples{
local_simes(c(0.01, 0.04, 0.10, 0.20))
}
