% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_adjusters.R
\name{local_unadj_all_ps}
\alias{local_unadj_all_ps}
\title{Unadjusted local step}
\usage{
local_unadj_all_ps(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha}
}
\value{
A single numeric value: just passes through the unadjusted p values
}
\description{
Given \eqn{k} child p-values just pass them through for simulation purposes
}
\details{
The idea is to not adjust so that we can compare the consequences of not adjusting to adjusting
}
