% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_scaled_beta.R
\name{r_scaled_beta}
\alias{r_scaled_beta}
\title{Sample from a Scaled Beta Distribution}
\usage{
r_scaled_beta(a, shape1, shape2, n = 1)
}
\arguments{
\item{a}{Numeric. The lower bound of the output interval. Must be between 0 and 1 (exclusive).}

\item{shape1}{Numeric. The first shape parameter of the Beta distribution.}

\item{shape2}{Numeric. The second shape parameter of the Beta distribution.}

\item{n}{Integer. The number of random samples to generate (default is 1).}
}
\value{
A numeric vector of length n containing the transformed samples.
}
\description{
Generates random samples from a Beta distribution and then linearly transforms
them to lie in the interval \eqn{[a, 1]}.
}
\examples{
r_scaled_beta(0.2, 2, 5, n = 10)
}
