% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_test_sims.R
\name{simulate_single_run}
\alias{simulate_single_run}
\title{Simulate a Single Top-Down Run of the Hierarchical Simes Procedure}
\usage{
simulate_single_run(
  tree_info,
  alpha,
  alt,
  beta_params,
  local_adjust_fn = local_simes
)
}
\arguments{
\item{alpha}{Numeric in \eqn{(0,1)}; the significance level threshold.}

\item{alt}{Logical vector of length \eqn{\sum_{d=0}^{l} k^d}, indicating which nodes
are non-null (\code{TRUE}) vs. null (\code{FALSE}).}

\item{beta_params}{Numeric vector of length 2 for the Beta distribution
parameters of alternative nodes (see \code{\link{draw_node_p_value}}).}

\item{local_adjust_fn}{A function that takes the child p-values of a node and returns a single value (like the min or max p-value) that can be used to decide on a local gating procedure. For now the default is the \code{local_simes} function.}
}
\value{
A logical scalar: \code{TRUE} if at least one null node was
\eqn{\le \alpha} (false rejection) in this single run, \code{FALSE} otherwise. TODO maybe track FDR later.
}
\description{
Performs one realization of the top-down testing with strict gating
(\eqn{p > \alpha} stops further testing) plus a local Simes gate for each parent.
TODO maybe this should just apply the local Simes gate rather than do the whole thing?

@param tree_info; a list object like that created by \code{\link{get_level_info}}.
}
\details{
\enumerate{
\item The root node (level 0) is tested unconditionally:
if \code{alt[1] = FALSE}, its p-value is \eqn{\mathrm{Unif}(0,1)};
if \code{alt[1] = TRUE}, it is a scaled Beta(\code{beta_params}).
\item If a node's final p-value \eqn{> \alpha}, we do not generate child p-values
(gating).
\item If \eqn{\le \alpha}, we generate the child's p-values.  Then we apply
a \code{\link{local_simes}} test on those \eqn{k} child p-values.
If Simes \eqn{> \alpha}, we forcibly set those child p-values above \eqn{\alpha}
to block deeper testing.
\item We track any null node with p-value \eqn{\le \alpha}; that is counted as
a "false rejection."
}
}
\examples{
info <- get_level_info(k = 2, l = 2)
alt_vec <- assign_alt(t = 0.5, tree_info = info)
simulate_single_run(tree_info = info, alpha = 0.05, alt = alt_vec, beta_params = c(0.1, 1))

}
\seealso{
\code{\link{simulate_many_runs}}
}
