% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{TreeTestSim-package}
\alias{TreeTestSim}
\alias{TreeTestSim-package}
\title{TreeTestSim: Enables simulation of hypothesis tests on tree-like structures}
\description{
This package helps researchers study hypothesis testing procedures when those hypotheses are organized in a tree-like structure. It has functions to generate k-ary trees with known proportions of leaves and nodes have non-zero or zero effects, enables the creation of simulated block or strata randomized experimental designs that can be organized in a nested tree-like format, and has helper functions which can be run to explore the impact of different analytic decisions on the outcomes of the testing.
}
\author{
\strong{Maintainer}: Jake Bowers \email{jake@jakebowers.org}

}
\keyword{internal}
